+++ example/a.ts
@@ -1,22 +1,18 @@
 import { Elysia, t } from '../src'
 import { z } from 'zod'
+import { req } from '../test/utils'
 
 const app = new Elysia()
-	.model({
-		sign: t.Object({
-			username: t.String(),
-			password: t.String()
-		}),
-		zodSign: z.object({
-			username: z.string(),
-			password: z.string()
-		})
+	.onError(() => {
+		// Condition 1: onError returns a plain object (not a Response instance)
+		return { hello: 'world' }
 	})
+	.mapResponse(({ responseValue }) => {
+		return new Response('A')
+	})
+	.get('/', () => 'hello')
+	.listen(3000)
 
-// For TypeBox
-// const zodSign = app.models.zodSign.Schema()
-// type zodSign = z.infer<typeof zodSign>
-//    ^?
-
-
-console.log(app.models.sign.schema)
+app.handle(req('/'))
+	.then(x => x.text())
+	.then(console.log)